+++ StripeIdentity/StripeIdentity/Source/NativeComponents/ViewControllers/DocumentCaptureViewController.swift
@@ -321,7 +321,9 @@ final class DocumentCaptureViewController: IdentityFlowViewController {
                 scanner: anyDocumentScanner,
                 concurrencyManager: concurrencyManager
                     ?? ImageScanningConcurrencyManager(
-                        sheetController: sheetController
+                        sheetController: sheetController,
+                        scannerName: .document,
+                        screenName: .documentCapture
                     ),
                 cameraPermissionsManager: cameraPermissionsManager,
                 appSettingsHelper: appSettingsHelper
@@ -482,7 +484,9 @@ extension DocumentCaptureViewController: ImageScanningSessionDelegate {
         }
         sheetController.analyticsClient.logCameraError(
             sheetController: sheetController,
-            error: error
+            error: error,
+            screenName: analyticsScreenName,
+            cameraSource: .cameraSession
         )
     }
 
@@ -495,7 +499,9 @@ extension DocumentCaptureViewController: ImageScanningSessionDelegate {
         }
         sheetController.analyticsClient.logCameraPermissionsChecked(
             sheetController: sheetController,
-            isGranted: isGranted
+            isGranted: isGranted,
+            screenName: analyticsScreenName,
+            cameraSource: .cameraSession
         )
     }