+++ StripeIdentity/StripeIdentity/Source/NativeComponents/ViewControllers/IdentityFlowViewController.swift
@@ -16,6 +16,7 @@ class IdentityFlowViewController: UIViewController {
     private let flowView = IdentityFlowView()
 
     private var navBarBackgroundColor: UIColor?
+    private var previousScreenNameForScreenAppearedAnalytic: IdentityAnalyticsClient.ScreenName?
 
     let analyticsScreenName: IdentityAnalyticsClient.ScreenName
 
@@ -71,6 +72,9 @@ class IdentityFlowViewController: UIViewController {
 
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
+        previousScreenNameForScreenAppearedAnalytic = (
+            transitionCoordinator?.viewController(forKey: .from) as? IdentityFlowViewController
+        )?.analyticsScreenName
         navigationController?.setNavigationBarBackgroundColor(with: navBarBackgroundColor)
     }
 
@@ -79,12 +83,15 @@ class IdentityFlowViewController: UIViewController {
         guard let sheetController = sheetController else {
             return
         }
+        let previousScreenName = previousScreenNameForScreenAppearedAnalytic
+        previousScreenNameForScreenAppearedAnalytic = nil
         sheetController.analyticsClient.stopTrackingTimeToScreenAndLogIfNeeded(
             to: analyticsScreenName,
             sheetController: sheetController
         )
         sheetController.analyticsClient.logScreenAppeared(
             screenName: analyticsScreenName,
+            previousScreenName: previousScreenName,
             sheetController: sheetController
         )
     }
@@ -100,7 +107,14 @@ class IdentityFlowViewController: UIViewController {
             try flowView.configure(with: viewModel)
         } catch {
             if let sheetController = sheetController {
-                sheetController.analyticsClient.logGenericError(error: error, sheetController: sheetController)
+                sheetController.analyticsClient.logGenericError(
+                    error: error,
+                    additionalMetadata: [
+                        "error_context": "flow_view_configure",
+                        "screen_name": analyticsScreenName.rawValue,
+                    ],
+                    sheetController: sheetController
+                )
             }
         }
         navBarBackgroundColor = viewModel.headerViewModel?.backgroundColor
@@ -169,6 +183,10 @@ extension IdentityFlowViewController {
         if let sheetController = self.sheetController {
             sheetController.analyticsClient.logGenericError(
                 error: BottomSheetError(loggableType: errorContent),
+                additionalMetadata: [
+                    "error_context": "bottomsheet",
+                    "screen_name": analyticsScreenName.rawValue,
+                ],
                 sheetController: sheetController
             )
         }